Review.
For now mostly syntactic and conventions.

I only changed some of the camlCase to kebab-case.

There are a lot of explicit `yield` calls. Not clear if all of them
  are needed.

Logging with `tags` is usually the preferred way.

String-interpolation is your friend. (Also calls `.stringify`
    automatically).

Missing toitdocs. They should be there at least for public elements.

A few too many abbreviatons.