comms: sendOpen in catch-and-restart
This avoids an exception bubbling all the way
to the top in the case that the comms are just not
ready yet for whatever reason, and allows some reries

We block on this sendOpen, so that we get a response
before continuing to process anything else comms
related, including sending heatbeats.

This was resulting in...
EXCEPTION error.
Failed to open device link
0: Comms.sendOpen_           <pkg:toit-lightbug>/services/comms/comms.toit:82:7
1: Comms.start_              <pkg:toit-lightbug>/services/comms/comms.toit:74:7
2: Comms                     <pkg:toit-lightbug>/services/comms/comms.toit:60:5