Use sequential ID generation for messages
Some cloud related processes are starting to make
some assumptions that we generally submit messages
with sequential message IDs, indicating either the
order that data was collected in, or the order it
perhaps should be used in.
So remove any hints of random ID generation, even if
custom apps might choose to use it for some reason,
we will stick to sequential.